requirements.txt
# Dino Dynasty OS Dependencies
pyyaml>=6.0
requests>=2.31.0
apscheduler>=3.10.0
